refactor: migrate vpc, vpc_subnet, vpc_router, vpc_router_route, vpc_internet_gateway to subpackages
Move VPC-related resources and datasources into their own packages
under internal/provider/resources/.

Amp-Thread: https://ampcode.com/threads/T-019d56b5-b859-72b6-9431-d6b3aebf9860
