+++ internal/provider/resources/vpc/datasource.go
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-package provider
+package vpc
 
 import (
 	"context"
@@ -18,20 +18,20 @@ import (
 )
 
 var (
-	_ datasource.DataSource              = (*vpcDataSource)(nil)
-	_ datasource.DataSourceWithConfigure = (*vpcDataSource)(nil)
+	_ datasource.DataSource              = (*DataSource)(nil)
+	_ datasource.DataSourceWithConfigure = (*DataSource)(nil)
 )
 
-// NewVPCDataSource initialises an images datasource
-func NewVPCDataSource() datasource.DataSource {
-	return &vpcDataSource{}
+// NewDataSource initialises a VPC datasource
+func NewDataSource() datasource.DataSource {
+	return &DataSource{}
 }
 
-type vpcDataSource struct {
+type DataSource struct {
 	client *oxide.Client
 }
 
-type vpcDataSourceModel struct {
+type DataSourceModel struct {
 	Description    types.String   `tfsdk:"description"`
 	DNSName        types.String   `tfsdk:"dns_name"`
 	ID             types.String   `tfsdk:"id"`
@@ -45,7 +45,7 @@ type vpcDataSourceModel struct {
 	Timeouts       timeouts.Value `tfsdk:"timeouts"`
 }
 
-func (d *vpcDataSource) Metadata(
+func (d *DataSource) Metadata(
 	ctx context.Context,
 	req datasource.MetadataRequest,
 	resp *datasource.MetadataResponse,
@@ -54,7 +54,7 @@ func (d *vpcDataSource) Metadata(
 }
 
 // Configure adds the provider configured client to the data source.
-func (d *vpcDataSource) Configure(
+func (d *DataSource) Configure(
 	_ context.Context,
 	req datasource.ConfigureRequest,
 	_ *datasource.ConfigureResponse,
@@ -66,7 +66,7 @@ func (d *vpcDataSource) Configure(
 	d.client = req.ProviderData.(*oxide.Client)
 }
 
-func (d *vpcDataSource) Schema(
+func (d *DataSource) Schema(
 	ctx context.Context,
 	req datasource.SchemaRequest,
 	resp *datasource.SchemaResponse,
@@ -121,12 +121,12 @@ Retrieve information about a specified VPC.
 	}
 }
 
-func (d *vpcDataSource) Read(
+func (d *DataSource) Read(
 	ctx context.Context,
 	req datasource.ReadRequest,
 	resp *datasource.ReadResponse,
 ) {
-	var state vpcDataSourceModel
+	var state DataSourceModel
 
 	// Read Terraform configuration data into the model
 	resp.Diagnostics.Append(req.Config.Get(ctx, &state)...)
