+++ internal/provider/resources/vpc_router_route/datasource_test.go
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-package provider
+package vpc_router_route_test
 
 import (
 	"fmt"
@@ -10,6 +10,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/oxidecomputer/oxide.go/oxide"
+	"github.com/oxidecomputer/terraform-provider-oxide/internal/provider"
 )
 
 type dataSourceVPCRouterRouteConfig struct {
@@ -30,11 +31,11 @@ data "oxide_vpc_router_route" "{{.BlockName}}" {
 `
 
 func TestAccCloudDataSourceVPCRouterRoute_full(t *testing.T) {
-	blockName := NewBlockName("datasource-vpc-router")
-	config, err := ParsedAccConfig(
+	blockName := provider.NewBlockName("datasource-vpc-router")
+	config, err := provider.ParsedAccConfig(
 		dataSourceVPCRouterRouteConfig{
 			BlockName:        blockName,
-			SupportBlockName: NewBlockName("support"),
+			SupportBlockName: provider.NewBlockName("support"),
 		},
 		dataSourceVPCRouterRouteConfigTpl,
 	)
@@ -43,8 +44,8 @@ func TestAccCloudDataSourceVPCRouterRoute_full(t *testing.T) {
 	}
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:                 func() { PreCheck(t) },
-		ProtoV6ProviderFactories: ProviderFactories(),
+		PreCheck:                 func() { provider.PreCheck(t) },
+		ProtoV6ProviderFactories: provider.ProviderFactories(),
 		Steps: []resource.TestStep{
 			{
 				Config: config,
