+++ internal/provider/resources/vpc_subnet/resource_test.go
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-package provider
+package vpc_subnet_test
 
 import (
 	"context"
@@ -13,6 +13,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/oxidecomputer/oxide.go/oxide"
+	"github.com/oxidecomputer/terraform-provider-oxide/internal/provider"
 	"github.com/oxidecomputer/terraform-provider-oxide/internal/provider/shared"
 )
 
@@ -53,9 +54,12 @@ resource "oxide_vpc_subnet" "test" {
 }
 `
 
-func buildVPCSubnetConfig(t *testing.T, cfg vpcSubnetTestConfig) string {
+func buildVPCSubnetConfig(
+	t *testing.T,
+	cfg vpcSubnetTestConfig,
+) string {
 	t.Helper()
-	config, err := ParsedAccConfig(cfg, vpcSubnetConfigTpl)
+	config, err := provider.ParsedAccConfig(cfg, vpcSubnetConfigTpl)
 	if err != nil {
 		t.Fatalf("error parsing config template: %v", err)
 	}
@@ -83,8 +87,8 @@ func TestAccCloudResourceVPCSubnet_full(t *testing.T) {
 	ipv6Config.IPv6Block = "fdfe:f6a5:5f06:a643::/64"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:                 func() { PreCheck(t) },
-		ProtoV6ProviderFactories: ProviderFactories(),
+		PreCheck:                 func() { provider.PreCheck(t) },
+		ProtoV6ProviderFactories: provider.ProviderFactories(),
 		CheckDestroy:             testAccVPCSubnetDestroy,
 		Steps: []resource.TestStep{
 			{
@@ -160,7 +164,7 @@ func checkResourceVPCSubnetIPv6(subnetName string) resource.TestCheckFunc {
 }
 
 func testAccVPCSubnetDestroy(s *terraform.State) error {
-	client, err := NewTestClient()
+	client, err := provider.NewTestClient()
 	if err != nil {
 		return err
 	}
